requirements.txt
snowflake-connector-python==3.7.1
snowflake-sqlalchemy==1.5.3
apache-airflow-providers-snowflake==5.5.2
apache-airflow-providers-microsoft-mssql==3.7.2